pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["dibkb <dibas9110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
unicorn = "^2.1.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
